Added archive calendar (using CSS/JS/Jquery); more PSR compliance
FullCalendar was chosen as the calendar for the archive.Nice-to-know jQuery script. Front-end is still quite raw though. Also, PSR-12 was applied to a lot more stuff but still not on the front-end, I would rather finish all (or at least most) code before undergoing the compliance process. We're going at a steady pace, I am glad.