Added recurrency!
There it is, the logic for recurrency! I am for now printing the event dates to the View, mostly as debugging. I have validated them and they're all correct. Now that we have got that out of our way, I guess our backend is mostly done - just needs some final touches here and there, mostly organizing. As for the front-end, changes will come on the next commit.