Finish archive page, compliance
This is neat! Our archive page now looks beautiful, shows the events as we want them (recurring in calendar), has a navigation option, and the whole template system was entirely restructured. As for the rest, I added an exclusive file for constants (complying with PSR-12), included the helpers.php on main.php (complying with PSR-12; helper  functions are now loaded and available everywhere in the plugin), lots of CSS adjustments, and began setting some ground for i18n.