MVP achieved: front \& backend functionally complete
So now we have our backend and frontend functionally complete. What we need to do now that we have a MVP is optimizations. Here's an interesting principle: every one-time-only event is in reality a recurrent event with a recurrency of 0 (aka it happens once and never again). This has some very interesting implications in our code, and opens up space for optimization. As for the rest, I must speed up that i18n, and then PSR-2 (2, not 12\!) compliance, then we're finished.